Move tools/device/modm -> tools/device/modm_devices
This prevents shadowing of the modm package from PyPi.
