+++ tools/device/modm_devices/resources/schema/device.xsd
